Battle Menu and Status Effect bugfix
Changes the Battle Menu to work as intended. The player can now choose actions by directional pad instead of hardcoded numbers. It now also loads the spells from the player's spell list instead of having a hardcoded spell.
This commit also implements changes to status effects and the battle scene so that each status effect can now tick before and/or after an action. It also fixes the bug where the status effect would get removed at the end of combat while the player was still iterating through the list to resolve actions.